+++ plone/app/contenttypes/migration/migration.py
@@ -467,13 +467,18 @@ def migrateCustomAT(fields_mapping, src_type, dst_type, dry_run=False):
     if fti is None or IDexterityFTI.providedBy(fti):
         # Get the needed info from an instance of the type
         catalog = portal.portal_catalog
-        brain = catalog(portal_type=src_type, sort_limit=1)[0]
-        src_obj = brain.getObject()
-        if IDexterityContent.providedBy(src_obj):
-            logger.error(
-                '%s should not be dexterity object!' % src_obj.absolute_url())
-        is_folderish = getattr(src_obj, 'isPrincipiaFolderish', False)
-        src_meta_type = src_obj.meta_type
+        brains = catalog(portal_type=src_type, sort_limit=1)
+        if not brains:
+            # no item? assume stuff
+            is_folderish = False
+            src_meta_type = src_type
+        else:
+            src_obj = brains[0].getObject()
+            if IDexterityContent.providedBy(src_obj):
+                logger.error(
+                    '%s should not be dexterity object!' % src_obj.absolute_url())
+            is_folderish = getattr(src_obj, 'isPrincipiaFolderish', False)
+            src_meta_type = src_obj.meta_type
     else:
         # Get info from at-fti
         src_meta_type = fti.content_meta_type