+++ plone/app/contenttypes/tests/test_migration_custom.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone.utils import safe_unicode
 from datetime import datetime
 from plone.app.contenttypes.migration.field_migrators import migrate_filefield
@@ -6,14 +7,20 @@ from plone.app.contenttypes.migration.field_migrators import migrate_imagefield
 from plone.app.contenttypes.migration.field_migrators import \
     migrate_simplefield
 from plone.app.contenttypes.migration.utils import installTypeIfNeeded
-from plone.app.contenttypes.testing import \
-    PLONE_APP_CONTENTTYPES_MIGRATION_TESTING
+from plone.app.contenttypes.testing import PLONE_APP_CONTENTTYPES_MIGRATION_FUNCTIONAL_TESTING  # noqa
+from plone.app.contenttypes.testing import PLONE_APP_CONTENTTYPES_MIGRATION_TESTING  # noqa
+from plone.app.testing import SITE_OWNER_NAME
+from plone.app.testing import SITE_OWNER_PASSWORD
 from plone.app.testing import TEST_USER_ID
 from plone.app.testing import applyProfile
+from plone.app.testing import applyProfile
+from plone.app.testing import login
 from plone.app.testing import setRoles
+from plone.testing.z2 import Browser
 
 import pytz
 import os.path
+import transaction
 import unittest2 as unittest
 
 
@@ -395,3 +402,123 @@ class MigrateCustomATTest(unittest.TestCase):
             os.environ['TZ'] = oldTZ
         else:
             del os.environ['TZ']
+
+
+class CustomMigrationFunctionalTests(unittest.TestCase):
+
+    layer = PLONE_APP_CONTENTTYPES_MIGRATION_FUNCTIONAL_TESTING
+
+    def setUp(self):
+        app = self.layer['app']
+        self.portal = self.layer['portal']
+        self.request = self.layer['request']
+        self.request['ACTUAL_URL'] = self.portal.absolute_url()
+        self.request['URL'] = self.portal.absolute_url()
+        self.catalog = getToolByName(self.portal, "portal_catalog")
+        self.portal.acl_users.userFolderAddUser('admin',
+                                                'secret',
+                                                ['Manager'],
+                                                [])
+        login(self.portal, 'admin')
+        self.portal.portal_workflow.setDefaultChain(
+            "simple_publication_workflow")
+        self.portal_url = self.portal.absolute_url()
+
+        self.browser = Browser(app)
+        self.browser.handleErrors = False
+        self.browser.addHeader(
+            'Authorization',
+            'Basic %s:%s' % (SITE_OWNER_NAME, SITE_OWNER_PASSWORD,)
+        )
+
+    def tearDown(self):
+        try:
+            applyProfile(self.portal, 'plone.app.contenttypes:uninstall')
+        except KeyError:
+            pass
+
+    def test_custom_migration_form(self):
+        """Basic test for the custom_migration form.
+        Field-mapping only works with javascript enabled so we migrate
+        only the content but not the fields.
+        """
+        # add some at content
+        self.portal.invokeFactory('Document', 'doc1')
+        self.portal.invokeFactory('Event', 'event1')
+        self.portal.event1.setTitle(u'Ein Törmin')
+        self.portal.event1.setDescription(u'Wänn?')
+        self.portal.doc1.setTitle(u'Ein Döcument')
+        self.portal.doc1.setDescription(u'Sö was')
+        transaction.commit()
+        self.browser.open('%s/@@pac_installer' % self.portal_url)
+        self.browser.getControl('Install').click()
+        # open custom-migration-form
+        self.browser.open('%s/@@custom_migration' % self.portal_url)
+        results = self.browser.contents
+        self.assertIn('Custom types migration control panel', results)
+        self.assertIn('<input type="hidden" name="Document:list" value="text__type__Products.Archetypes.Field.TextField" />', results)  # noqa
+        self.assertEqual(self.browser.getControl(name='dx_select_Document').value, [''])  # noqa
+        # chose to migrate to Link
+        self.browser.getControl(name='dx_select_Document').value = ['Link']
+        self.assertIn('<input type="hidden" name="Event:list" value="startDate__type__Products.Archetypes.Field.DateTimeField" />', results)  # noqa
+        # chose to migrate to Link
+        self.browser.getControl(name='dx_select_Event').value = ['Link']
+        # run migration
+        self.browser.getControl(name='form.button.Migrate').click()
+        results = self.browser.contents
+        self.assertIn('Migration applied succesfully for 1 "Document" items.', results)  # noqa
+        self.assertIn('Migration applied succesfully for 1 "Event" items.', results)  # noqa
+        self.assertIn('No content to migrate.', results)
+        link1 = self.portal['doc1']
+        self.assertEqual(link1.portal_type, 'Link')
+        self.assertEqual(link1.title, u'Ein D\xf6cument')
+        self.assertEqual(link1.description, u'S\xf6 was')
+        self.assertEqual(self.portal['event1'].portal_type, 'Link')
+        # we did not migrate the fields so lets find out if it is a real Link
+        link1.remote_url = 'http://www.starzel.de'
+        view = link1()
+        self.assertIn(u'<h1 class="documentFirstHeading">Ein D\xf6cument</h1>', view)  # noqa
+        self.assertIn(u'The link address is:</span>\n            <a href="http://www.starzel.de">http://www.starzel.de</a>', view)  # noqa
+
+    def test_custom_migration_form_for_types_with_spaces(self):
+        """Basic test for the custom_migration form.
+        Field-mapping only works with javascript enabled so we migrate
+        only the content but not the fields.
+        """
+        # add some at content
+        self.portal.invokeFactory('News Item', 'news1')
+        self.portal.invokeFactory('Event', 'event1')
+        self.portal.event1.setTitle(u'Ein Törmin')
+        self.portal.event1.setDescription(u'Wänn?')
+        self.portal.news1.setTitle(u'Ein News Item')
+        self.portal.news1.setDescription(u'Sö was')
+        transaction.commit()
+        self.browser.open('%s/@@pac_installer' % self.portal_url)
+        self.browser.getControl('Install').click()
+        # open custom-migration-form
+        self.browser.open('%s/@@custom_migration' % self.portal_url)
+        results = self.browser.contents
+        self.assertIn('Custom types migration control panel', results)
+        self.assertIn('<input type="hidden" name="News_space_Item:list" value="text__type__Products.Archetypes.Field.TextField" />', results)  # noqa
+        self.assertEqual(self.browser.getControl(name='dx_select_News_space_Item').value, [''])  # noqa
+        # chose to migrate to Link
+        self.browser.getControl(name='dx_select_News_space_Item').value = ['Link']  # noqa
+        self.assertIn('<input type="hidden" name="Event:list" value="startDate__type__Products.Archetypes.Field.DateTimeField" />', results)  # noqa
+        # chose to migrate to Link
+        self.browser.getControl(name='dx_select_Event').value = ['Link']
+        # run migration
+        self.browser.getControl(name='form.button.Migrate').click()
+        results = self.browser.contents
+        self.assertIn('Migration applied succesfully for 1 "News Item" items.', results)  # noqa
+        self.assertIn('Migration applied succesfully for 1 "Event" items.', results)  # noqa
+        self.assertIn('No content to migrate.', results)
+        link1 = self.portal['news1']
+        self.assertEqual(link1.portal_type, 'Link')
+        self.assertEqual(link1.title, u'Ein News Item')
+        self.assertEqual(link1.description, u'S\xf6 was')
+        self.assertEqual(self.portal['event1'].portal_type, 'Link')
+        # we did not migrate the fields so lets find out if it is a real Link
+        link1.remote_url = 'http://www.starzel.de'
+        view = link1()
+        self.assertIn(u'<h1 class="documentFirstHeading">Ein News Item</h1>', view)  # noqa
+        self.assertIn(u'The link address is:</span>\n            <a href="http://www.starzel.de">http://www.starzel.de</a>', view)  # noqa